fixed country popup, removed unused components
